v2.4.9.7 -> v2.4.9.8
  - Christoph Hellwig: clean up personality handling a bit
  - Robert Love: update sysctl/vm documentation
  - make the three-argument (that everybody hates) "min()" be "min_t()",
  and introduce a type-anal "min()" that complains about arguments of
  different types.

+++ fs/ncpfs/ioctl.c
@@ -378,7 +378,7 @@ outrel:
 		}
 		{
 			struct ncp_objectname_ioctl user;
-			int outl;
+			size_t outl;
 
 			if (copy_from_user(&user, 
 					   (struct ncp_objectname_ioctl*)arg,
@@ -448,7 +448,7 @@ outrel:
 		}
 		{
 			struct ncp_privatedata_ioctl user;
-			int outl;
+			size_t outl;
 
 			if (copy_from_user(&user, 
 					   (struct ncp_privatedata_ioctl*)arg,
